requirements.txt
strands-agents
mcp>=1.9.4
uv>=0.7.13
requests>=2.31.0
python-dotenv